// src/generated/api/models/billSplitServiceApplicationFeaturesBillsListBillsListBillsHandlerItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Bill Split Service API
 * API for bill splitting, Google login, payment tracking and administration.
 * OpenAPI spec version: v1
 */

export interface BillSplitServiceApplicationFeaturesBillsListBillsListBillsHandlerItem {
  billId?: string;
  /** @nullable */
  title?: string | null;
  totalAmount?: number;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  dueDate?: string | null;
  /** @nullable */
  assignedOrCollectedAmount?: number | null;
  /** @nullable */
  remainingAmount?: number | null;
}